Break up index.ts
